requirements.txt
asn1crypto==1.5.1
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
ckanapi==4.8
docopt==0.6.2
greenlet==3.2.3
idna==3.10
lxml==6.0.0
pg8000==1.31.4
psycopg2==2.9.8
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
requests==2.32.4
scramp==1.4.6
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
text-unidecode==1.3
typing_extensions==4.14.1
urllib3==2.5.0
